Introduce OtherSamples.kt to generate tables for writerside docs outside Korro
