+++ core/generated-sources/src/test/kotlin/org/jetbrains/kotlinx/dataframe/explainer/PluginCallbackProxy.kt
@@ -75,32 +75,7 @@ object PluginCallbackProxy : PluginCallback {
         var output: DataFrameHtmlData
         val manualOutput = this.manualOutput
         if (manualOutput == null) {
-            output = DataFrameHtmlData.tableDefinitions() + DataFrameHtmlData(
-                // copy writerside stlyles
-                style = """
-                    body {
-                        font-family: "JetBrains Mono",SFMono-Regular,Consolas,"Liberation Mono",Menlo,Courier,monospace;
-                    }       
-                    
-                    :root {
-                        color: #19191C;
-                        background-color: #fff;
-                    }
-                    
-                    :root[theme="dark"] {
-                        background-color: #19191C;
-                        color: #FFFFFFCC
-                    }
-                    
-                    details details {
-                        margin-left: 20px; 
-                    }
-                    
-                    summary {
-                        padding: 6px;
-                    }
-                """.trimIndent()
-            )
+            output = DataFrameHtmlData.tableDefinitions() + WritersideStyle
 
             // make copy to avoid concurrent modification exception
             val statements = expressionsByStatement.toMap()
@@ -243,7 +218,7 @@ object PluginCallbackProxy : PluginCallback {
 }
 
 private fun convertToHTML(dataframeLike: Any): DataFrameHtmlData {
-    fun DataFrame<*>.toHTML() = toHTML(SamplesDisplayConfiguration, getFooter = { "" })
+    fun DataFrame<*>.toHTML() = toHTML(SamplesDisplayConfiguration, getFooter = WritersideFooter)
     fun FormattedFrame<*>.toHTML1() = toHTML(SamplesDisplayConfiguration)
 
     return when (dataframeLike) {
