--- core/src/test/kotlin/org/jetbrains/kotlinx/dataframe/explainer/SamplesDisplayConfiguration.kt
@@ -1,5 +1,0 @@
-package org.jetbrains.kotlinx.dataframe.explainer
-
-import org.jetbrains.kotlinx.dataframe.io.DisplayConfiguration
-
-val SamplesDisplayConfiguration = DisplayConfiguration(enableFallbackStaticTables = false)
